--- js/reduce.algo2/algo2_tests.ts.md
@@ -1,50 +1,0 @@
-# Snapshot report for `reduce.algo2/algo2_tests.ts`
-
-The actual snapshot is saved in `algo2_tests.ts.snap`.
-
-Generated by [AVA](https://avajs.dev).
-
-## algo2.map() retourne toutes les propriétés d'entreprise attendues par l'algo d'apprentissage
-
-> Snapshot 1
-
-    [
-      {
-        _id: {
-          batch: '1910',
-          periode: Date 2015-12-01 00:00:00 UTC {},
-          siren: '012345678',
-          type: 'other',
-        },
-        value: {
-          entreprise: {
-            arrete_bilan_bdf: Date 2015-04-28 00:00:00 UTC {},
-            exercice_bdf: 0,
-            paydex_nb_jours: 1,
-            paydex_nb_jours_past_1: null,
-            paydex_nb_jours_past_12: null,
-            periode: Date 2015-12-01 00:00:00 UTC {},
-            siren: '012345678',
-          },
-        },
-      },
-      {
-        _id: {
-          batch: '1910',
-          periode: Date 2016-01-01 00:00:00 UTC {},
-          siren: '012345678',
-          type: 'other',
-        },
-        value: {
-          entreprise: {
-            arrete_bilan_bdf: Date 2015-04-28 00:00:00 UTC {},
-            exercice_bdf: 0,
-            paydex_nb_jours: 2,
-            paydex_nb_jours_past_1: 1,
-            paydex_nb_jours_past_12: null,
-            periode: Date 2016-01-01 00:00:00 UTC {},
-            siren: '012345678',
-          },
-        },
-      },
-    ]
